+++ commonmark/src/main/java/org/commonmark/renderer/text/TextContentWriter.java
@@ -8,6 +8,7 @@ public class TextContentWriter {
     private final Appendable buffer;
     private final LineBreakRendering lineBreakRendering;
 
+    private final LinkedList<String> prefixes = new LinkedList<>();
     private final LinkedList<Boolean> tight = new LinkedList<>();
 
     private String blockSeparator = null;
@@ -36,6 +37,7 @@ public class TextContentWriter {
 
     public void line() {
         append('\n');
+        writePrefixes();
     }
 
     public void block() {
@@ -62,6 +64,32 @@ public class TextContentWriter {
     }
 
     /**
+     * Push a prefix onto the top of the stack. All prefixes are written at the beginning of each line, until the
+     * prefix is popped again.
+     *
+     * @param prefix the raw prefix string
+     */
+    public void pushPrefix(String prefix) {
+        prefixes.addLast(prefix);
+    }
+
+    /**
+     * Write a prefix.
+     *
+     * @param prefix the raw prefix string to write
+     */
+    public void writePrefix(String prefix) {
+        write(prefix);
+    }
+
+    /**
+     * Remove the last prefix from the top of the stack.
+     */
+    public void popPrefix() {
+        prefixes.removeLast();
+    }
+
+    /**
      * Change whether blocks are tight or loose. Loose is the default where blocks are separated by a blank line. Tight
      * is where blocks are not separated by a blank line. Tight blocks are used in lists, if there are no blank lines
      * within the list.
@@ -84,12 +112,26 @@ public class TextContentWriter {
         return !tight.isEmpty() && tight.getLast();
     }
 
+    private void writePrefixes() {
+        for (String prefix : prefixes) {
+            append(prefix);
+        }
+    }
+
     /**
      * If a block separator has been enqueued with {@link #block()} but not yet written, write it now.
      */
     private void flushBlockSeparator() {
         if (blockSeparator != null) {
-            append(blockSeparator);
+            if (blockSeparator.equals("\n") || blockSeparator.equals("\n\n")) {
+                for (int i = 0; i < blockSeparator.length(); i++) {
+                    var sep = blockSeparator.charAt(i);
+                    append(sep);
+                    writePrefixes();
+                }
+            } else {
+                append(blockSeparator);
+            }
             blockSeparator = null;
         }
     }